requirements.txt
python-dotenv==1.0.1
requests==2.32.3
backoff==2.2.1
openai==1.51.0
numpy==2.1.1
scikit-learn==1.5.2
orjson==3.10.7
uvloop==0.20.0; platform_system != "Windows"
